TIKA-3266 -- enable user-configured OCR parsers
