+++ tika-parsers/tika-parsers-classic/tika-parsers-classic-modules/tika-parser-image-module/src/main/java/org/apache/tika/parser/image/JpegParser.java
@@ -58,6 +58,5 @@ public class JpegParser extends AbstractImageParser {
         } finally {
             tmp.dispose();
         }
-
     }
 }
