+++ tika-parsers/tika-parsers-classic/tika-parsers-classic-modules/tika-parser-pdf-module/src/test/java/org/apache/tika/parser/pdf/PDFParserTest.java
@@ -37,7 +37,6 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import org.apache.commons.io.IOUtils;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.log4j.Level;
 import org.apache.log4j.Logger;
 import org.apache.pdfbox.rendering.ImageType;
@@ -66,8 +65,6 @@ import org.apache.tika.parser.CompositeParser;
 import org.apache.tika.parser.ParseContext;
 import org.apache.tika.parser.Parser;
 import org.apache.tika.parser.PasswordProvider;
-import org.apache.tika.parser.ocr.TesseractOCRConfig;
-import org.apache.tika.parser.ocr.TesseractOCRParser;
 import org.apache.tika.sax.AbstractRecursiveParserWrapperHandler;
 import org.apache.tika.sax.BodyContentHandler;
 import org.apache.tika.sax.ContentHandlerDecorator;
@@ -84,16 +81,6 @@ import org.xml.sax.ContentHandler;
 public class PDFParserTest extends TikaTest {
 
     public static Level PDFBOX_LOG_LEVEL = Level.INFO;
-
-    private static Boolean hasTesseract = null;
-
-    public static boolean canRunOCR() {
-        if (hasTesseract != null) {
-            return hasTesseract;
-        }
-        hasTesseract = new TesseractOCRParser().hasTesseract(new TesseractOCRConfig());
-        return hasTesseract;
-    }
 
     @BeforeClass
     public static void setup() {
@@ -586,37 +573,6 @@ public class PDFParserTest extends TikaTest {
         assertEquals("false", m.get(PDF.HAS_XFA));
         assertTrue("found", (result.xml.contains("<li>aTextField: TIKA-1226</li>")));
     }
-
-    @Test // TIKA-2232
-    public void testEmbeddedJBIG2Image() throws Exception {
-
-        ParseContext context = new ParseContext();
-        PDFParserConfig config = new PDFParserConfig();
-        config.setExtractInlineImages(true);
-        config.setExtractUniqueInlineImagesOnly(false);
-        context.set(PDFParserConfig.class, config);
-
-
-        List<Metadata> metadatas = getRecursiveMetadata("testPDF_JBIG2.pdf", context);
-        assertEquals(2, metadatas.size());
-        assertContains("test images compressed using JBIG2", metadatas.get(0).get(AbstractRecursiveParserWrapperHandler.TIKA_CONTENT));
-
-        for (String key : metadatas.get(1).names()) {
-            if (key.startsWith("X-TIKA:EXCEPTION")) {
-                fail("Exception: " + metadatas.get(1).get(key));
-            }
-        }
-        assertEquals("Invalid height.", "91", metadatas.get(1).get("height"));
-        assertEquals("Invalid width.", "352", metadatas.get(1).get("width"));
-        
-        assertNull(metadatas.get(0).get(TikaCoreProperties.RESOURCE_NAME_KEY));
-        assertEquals("image0.jb2",
-                metadatas.get(1).get(TikaCoreProperties.RESOURCE_NAME_KEY));
-        assertEquals(MediaType.image("x-jbig2").toString(), 
-                metadatas.get(1).get(Metadata.CONTENT_TYPE));
-    }
-
-
 
     @Test
     public void testSingleCloseDoc() throws Exception {
@@ -1116,76 +1072,6 @@ public class PDFParserTest extends TikaTest {
         assertTrue("Should have thrown exception", tikaEx);
         assertEquals(0, m.getValues(TikaCoreProperties.TIKA_META_EXCEPTION_WARNING).length);
         assertNotContained("1309.61", content);
-    }
-
-    @Test
-    public void testEmbeddedJPEG() throws Exception {
-        //TIKA-1990, test that an embedded jpeg is correctly decoded
-        PDFParserConfig config = new PDFParserConfig();
-        config.setExtractInlineImages(true);
-        ParseContext context = new ParseContext();
-        context.set(PDFParserConfig.class, config);
-
-        List<Metadata> metadataList = getRecursiveMetadata("testPDF_childAttachments.pdf", context);
-        //sanity check
-        assertEquals(5, metadataList.size());
-        //inlined jpeg metadata
-        Metadata jpegMetadata = metadataList.get(1);
-        assertEquals("image/jpeg", jpegMetadata.get(Metadata.CONTENT_TYPE));
-        //the metadata parse will fail if the stream is not correctly decoded
-        assertEquals("1425", jpegMetadata.get(Metadata.IMAGE_LENGTH));
-    }
-
-
-    @Test
-    public void testJBIG2OCROnly() throws Exception {
-        assumeTrue("can run OCR", canRunOCR());
-        PDFParserConfig config = new PDFParserConfig();
-        config.setOcrStrategy(PDFParserConfig.OCR_STRATEGY.OCR_ONLY);
-        ParseContext context = new ParseContext();
-        context.set(PDFParserConfig.class, config);
-        //make sure everything works with regular xml _and_ with recursive
-        XMLResult xmlResult = getXML("testPDF_JBIG2.pdf", context);
-        assertContains("Norconex", xmlResult.xml);
-    }
-
-    @Test
-    public void testOCRAutoMode() throws Exception {
-        assumeTrue("can run OCR", canRunOCR());
-        PDFParserConfig config = new PDFParserConfig();
-        config.setOcrStrategy(PDFParserConfig.OCR_STRATEGY.AUTO);
-        ParseContext context = new ParseContext();
-        context.set(PDFParserConfig.class, config);
-        XMLResult xmlResult = getXML("testOCR.pdf", context);
-        assertContains("Happy New Year", xmlResult.xml);
-
-        config.setOcrStrategy(PDFParserConfig.OCR_STRATEGY.NO_OCR);
-        String txt = getText("testOCR.pdf", new Metadata(), context);
-        assertEquals("", txt.trim());
-    }
-
-    @Test
-    public void testTesseractInitializationWorks() throws Exception {
-        //TIKA-2970 -- make sure that configurations set on the TesseractOCRParser
-        //make it through to when the TesseractOCRParser is called via
-        //the PDFParser
-        assumeTrue("can run OCR", canRunOCR());
-
-        //via the config, tesseract should skip this file because it is too large
-        try (InputStream is = getResourceAsStream("/org/apache/tika/parser/pdf/tika-ocr-config.xml")) {
-            assertNotNull(is);
-            TikaConfig tikaConfig = new TikaConfig(is);
-            Parser p = new AutoDetectParser(tikaConfig);
-            String text = getText(getResourceAsStream("/test-documents/testOCR.pdf"), p);
-            assertTrue(StringUtils.isAllBlank(text));
-
-            //now override the max file size to ocr, and you should get text
-            ParseContext pc = new ParseContext();
-            TesseractOCRConfig tesseractOCRConfig = new TesseractOCRConfig();
-            pc.set(TesseractOCRConfig.class, tesseractOCRConfig);
-            text = getText(getResourceAsStream("/test-documents/testOCR.pdf"), p, pc);
-            assertContains("Happy", text);
-        }
     }
 
     @Test
